Restructure views
