+++ BluetoothChargeCheck.UI/Popups/DeviceInfoPopup.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="DCT.BluetoothChargeCheck.Views.BatteryTrayIcon.DeviceInfoPopup"
+﻿<UserControl x:Class="DCT.BluetoothChargeCheck.Popups.DeviceInfoPopup"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
