+++ BluetoothChargeCheck.UI/Popups/MainPopup.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="DCT.BluetoothChargeCheck.Views.AppTrayIcon.MainPopup"
+﻿<UserControl x:Class="DCT.BluetoothChargeCheck.Popups.MainPopup"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
